Cartography Package

The cartography package allows various cartographic representations such as proportional symbols, chroropleth, typology, flows or discontinuities. In addition it also proposes some useful features like cartographic palettes, layout (scale, north arrow, title…), labels, legends or access to cartographic API to ease the graphic presentation of maps.

A vignette contains commented scripts on how to build various types of maps with cartography: vignette(topic = "cartography")

Main functions :

Proportional symbols maps (circles, squares, bars) propSymbolsLayer, propSymbolsChoroLayer, propSymbolsTypoLayer, propTrianglesLayer

Chroropleth maps (main discretization methods are available) choroLayer

Typology maps typoLayer

Flow maps (proportional and classified links) getLinkLayer, propLinkLayer, gradLinkLayer

Discontinuities maps (variable size and color of borders) getBorders, discLayer

Cartographic palettes (palettes adapted to cartographic representation) carto.pal

Layout (scale, north arrow, title...) layoutLayer

Labels labelLayer

Legends legendBarsSymbols, legendChoro, legendCirclesSymbols, legendGradLines, legendPropLines, legendPropTriangles, legendSquaresSymbols, legendTypo

Access to cartographic API (via rosm package) getTiles, tilesLayer

Irregular polygons to regular grid transformation with data handling getGridLayer, getGridData