Dot Density Layer

Plot a dot density layer.

dotDensityLayer(
  x,
  spdf,
  df,
  spdfid = NULL,
  dfid = NULL,
  var,
  n = NULL,
  pch = 1,
  cex = 0.15,
  type = "random",
  col = "black",
  legend.pos = "topright",
  legend.txt = NULL,
  legend.cex = 0.6,
  legend.col = "black",
  legend.frame = TRUE,
  add = TRUE
)

Arguments

x

an sf object, a simple feature collection. If x is used then spdf, df, spdfid and dfid are not.

spdf

a SpatialPolygonsDataFrame.

df

a data frame that contains the values to plot. If df is missing spdf@data is used instead.

spdfid

name of the identifier variable in spdf, default to the first column of the spdf data frame. (optional)

dfid

name of the identifier variable in df, default to the first column of df. (optional)

var

name of the numeric variable to plot.

n

one dot on the map represents n (in var units).

pch

symbol to use: points.

cex

size of the symbols

type

points allocation method: "random" or "regular" (see Details).

col

color of the points.

legend.pos

"topright", "left", "right", "bottomleft", "bottom", "bottomright". If legend.pos is "n" then the legend is not plotted.

legend.txt

text in the legend.

legend.cex

size of the legend text.

legend.col

color of the text in the legend.

legend.frame

whether to add a frame to the legend (TRUE) or not (FALSE).

add

whether to add the layer to an existing plot (TRUE) or not (FALSE).

Details

The type parameters is defined within the st_sample function.

See also

Examples

if (FALSE) {
library(sf)
mtq <- st_read(system.file("gpkg/mtq.gpkg", package="cartography"))
plot(st_geometry(mtq), col = "#B8704D50")
dotDensityLayer(x = mtq,  var="POP", pch=20, col = "red4", n = 200)
layoutLayer(title = "Population Distribution in Martinique, 2015")
}