Cartography Package

This package helps to design cartographic representations such as proportional symbols, choropleth, typology, flows or discontinuities maps. It also offers several features that improve the graphic presentation of maps, for instance, map palettes, layout elements (scale, north arrow, title...), labels or legends.

A vignette contains commented scripts on how to create various maps:
vignette(topic = "cartography", package = "cartography")
A cheat sheet displays a quick overview of cartography's main features:
vignette(topic = "cheatsheet" , package = "cartography")

Consider mapsf

There are no plans for new features or enhancements in cartography. Basic maintenance and support will continue indefinitely. Existing projects that use cartography can safely continue to use cartography. The mapsf R package is the successor of cartography and it is friendlier, lighter and more robust. See mapsf vignette (https://riatelab.github.io/mapsf/articles/mapsf.html) to migrate from cartography to mapsf.

Symbology

These functions build cartographic layers.

Proportional symbols layer (propSymbolsLayer, propSymbolsChoroLayer, propSymbolsTypoLayer, propTrianglesLayer)

Choropleth layer (choroLayer)

Typology layer (typoLayer)

Flows layer (propLinkLayer, gradLinkLayer, gradLinkTypoLayer)

Discontinuities layer (discLayer)

Access to cartographic OpenStreetMap APIs (tilesLayer)

Dot density layer (tilesLayer)

Labels layer (labelLayer)

Smooth layer (smoothLayer)

png layer (pngLayer)

Hatched/pattern layer (hatchedLayer)

Tag (or word) cloud layer (wordcloudLayer)

Transformations

These functions transform or create spatial objects.

Extract polygons borders (getBorders, getOuterBorders)

Build a regular grid layer (getGridLayer)

Create a links layer from a data.frame of links (getLinkLayer)

Create a png layer (getPngLayer)

Create a pencil layer (getPencilLayer)

Get tiles from OpenStreetMap servers (getTiles)

Map Layout

These functions are dedicated to the map layout design.

Scale bar (barscale)

Get figure dimensions (getFigDim)

Layout layer (layoutLayer)

North arrow (north)

Color Palettes

Use these function to build custom color palettes.

Build cartographic palettes (carto.pal)

Legends

These functions create legends.

Legends (legendBarsSymbols, legendChoro, legendCirclesSymbols, legendGradLines, legendPropLines, legendPropTriangles, legendSquaresSymbols, legendTypo, legendHatched)

Classification

This function uses classification methods for binning data.

Classification (getBreaks)